Fixing parsing bug using POSIX echo
If you were to use `echo "aaa\nbbb"`, you always get two lines of
output:

    aaa
    bbb

POSIX echo does not support any flags.  This is bad because the script
uses echo to feed information into sed and other commands.  The patterns
for seeking functions and variables are bound to the beginning of the
line, but that doesn't help when you have a line like

    sed -i 's/^\[Unit\]/&\nWants=something'

because the echo turns that into two lines

    sed -i 's/^\[Unit\]/&
    Wants=something'

and the second line appears to be setting a variable.  The other bits of
the code still treat this as one line, which is why any comment above
this type of line would trigger some documentation.