Updating the function matching pattern
This is based on testing Bash to see what it allows as function names.
Turns out Bash is quite impressive.  This is valid, and you're welcome
to try it out for yourself.

    [☃_{}() { echo "Valid function."; }
    [☃_{}

This patch closes issue #8 - thanks @phoccer for reporting this and for
sharing the link to Google's shell style guide.